% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abSamples.R
\name{abSamples}
\alias{abSamples}
\title{abSamples}
\usage{
abSamples(readsMatrix, conta = "auto", q = 0.99)
}
\arguments{
\item{readsMatrix}{a matrix of the number of reads per target}

\item{conta}{a parameter for the isotree function}

\item{q}{quantile}
}
\value{
the aberant and normal samples
}
\description{
abSamples
}
\examples{
readsMatrix = matrix(runif(500),nrow=50,ncol=10)
abSamples(readsMatrix)

}
