% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abTargets.R
\name{abTargets}
\alias{abTargets}
\title{abTargets}
\usage{
abTargets(readsMatrix, abSamples)
}
\arguments{
\item{readsMatrix}{the reads matrix}

\item{abSamples}{the abSamples list result of the abSamples function}
}
\value{
a list of dataframes of the targets tagged as outliers
}
\description{
abTargets
}
\examples{
abTargets(0,0)
}
