% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRatio.R
\name{calculateRatio}
\alias{calculateRatio}
\title{calculate Ratio}
\usage{
calculateRatio(abTargets, readsMatrix, abSamples)
}
\arguments{
\item{abTargets}{a list of dataframes of the targets tagged as outliers result of the abTargets() function}

\item{readsMatrix}{the reads matrix}

\item{abSamples}{the abSamples list result of the abSamples function}
}
\value{
the ratio associated with the abTargets
}
\description{
calculate Ratio
}
\examples{
calculateRatio(abTargets(readsMatrixExample,abSamples(readsMatrixExample)), readsMatrixExample,
abSamples(readsMatrixExample))
}
