% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createReadsMatrix.R
\name{CreateReadsMatrix}
\alias{CreateReadsMatrix}
\title{CreateReadsMatrix}
\usage{
CreateReadsMatrix(bamPath, bedFile, bedtools, outputFile = "n")
}
\arguments{
\item{bamPath}{a path leading to the .bam and .bai files}

\item{bedFile}{a path leading to the .bed file}

\item{bedtools}{the path leading to bedtools}

\item{outputFile}{(optional) a path leading to a text file}
}
\value{
a reads matrix
}
\description{
CreateReadsMatrix
}
\examples{
bamPath <- "/Users/admin/Documents/GitHub/ifCNVR/ifCNVR/inst/extdata/"
bed <- "/Users/admin/Documents/GitHub/ifCNVR/ifCNVR/inst/bedFile.bed"
bedtools <- "/Users/admin/miniconda3/bin/bedtools"
readsMatrix <- CreateReadsMatrix(bamPath, bed, bedtools)

}
