% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readsMatrixExample.R
\docType{data}
\name{readsMatrixExample}
\alias{readsMatrixExample}
\title{Example dataset}
\format{
A data frame with 11 rows and 3 variables:
\describe{
\item{targets}{the targets}
\item{sample_1}{the number of reads in sample 1}
\item{sample_2}{the number of reads in sample 2}
...
}
}
\source{
\url{EDF R\&D}
}
\usage{
readsMatrixExample
}
\description{
A dataset for the examples
}
\keyword{datasets}
