% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{scoring}
\alias{scoring}
\title{scoring}
\usage{
scoring(k, n, N)
}
\arguments{
\item{k}{number of modified targets on the region}

\item{n}{number of targets on the region}

\item{N}{number of targets in the panel}
}
\value{
the confidence score
}
\description{
scoring
}
\examples{
scoring(10,20,150)
}
