% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abTargets.R
\name{abTargets}
\alias{abTargets}
\title{abTargets}
\usage{
abTargets(readsMatrix, abSamples, opt = "regular", pred = 0.6)
}
\arguments{
\item{readsMatrix}{the reads matrix}

\item{abSamples}{the abSamples list result of the abSamples function}

\item{opt}{"regular" or "extensive" a parameter}

\item{pred}{a threshold on the isolation forest outlier prediction (range=[0,1[)}
}
\value{
a list of dataframes of the targets tagged as outliers
}
\description{
abTargets
}
\examples{
abTargets(readsMatrixExample,abSamples(readsMatrixExample))
}
