% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRatio.R
\name{calculateRatio}
\alias{calculateRatio}
\title{calculate Ratio}
\usage{
calculateRatio(readsMatrix, abSamples, roi, soi)
}
\arguments{
\item{readsMatrix}{the reads matrix}

\item{abSamples}{the abSamples list result of the abSamples function}

\item{roi}{the region of interest}

\item{soi}{the sample of interest}
}
\value{
the ratio associated with the abTargets
}
\description{
calculate Ratio
}
\examples{
calculateRatio(readsMatrixExample, abSamples(readsMatrixExample), "EGFR-Ex20", "sample_2")
}
