% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateScore.R
\name{calculateScore}
\alias{calculateScore}
\title{calculate Score}
\usage{
calculateScore(abTargets, bedFile, roi = "Gene-Exon", sep = "-", column = 4)
}
\arguments{
\item{abTargets}{a list of dataframes of the targets tagged as outliers result of the abTargets() function}

\item{bedFile}{a path leading to the .bed file}

\item{roi}{the region of interest (Gene or Gene-Exon)}

\item{sep}{a character the separator between roi in the bed file}

\item{column}{the column of the bed file with the roi}
}
\value{
the score associated with the abSamples in the desired roi
}
\description{
calculate Score
}
\examples{
bed <- "/Users/admin/Documents/GitHub/ifCNVR/ifCNVR/inst/bedFile.bed"
calculateScore(abTargets(readsMatrixExample,abSamples(readsMatrixExample)), bed)
}
