% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createReadsMatrix.R
\name{CreateReadsMatrix}
\alias{CreateReadsMatrix}
\title{CreateReadsMatrix}
\usage{
CreateReadsMatrix(
  bamPath,
  bedFile,
  bedtoolsPath,
  outputFile = "n",
  verbose = TRUE
)
}
\arguments{
\item{bamPath}{a path leading to the .bam and .bai files}

\item{bedFile}{a path leading to the .bed file (Warning replace chrX by X in the position)}

\item{bedtoolsPath}{the path leading to bedtools}

\item{outputFile}{(optional) a path leading to a text file}

\item{verbose}{a boolean}
}
\value{
a reads matrix
}
\description{
CreateReadsMatrix
}
\examples{
bamPath <- system.file("extdata/",package = "ifCNVR")
bed <- system.file("bedFile.bed",package = "ifCNVR")
bedtools <- 'n'
readsMatrix <- CreateReadsMatrix(bamPath, bed, bedtools)

}
