% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateReport.R
\name{generateReport}
\alias{generateReport}
\title{generateReport}
\usage{
generateReport(outputFile = "n")
}
\arguments{
\item{outputFile}{}
}
\value{

}
\description{
generateReport
}
\examples{
generateReport()
}
